criar declaracao builder
